Movve service definitions to provider
